Alt endelig fucking virker jesus christ. Json mangler stadig

+++ The Movies/ViewModel/CinemaViewModel.cs
@@ -12,40 +12,45 @@ namespace The_Movies.ViewModel
 {
     public class CinemaViewModel : ViewModelBase
     {
-
         private readonly ICinemaRepository _repo;
 
-        public ICommand AddCinemaCommand { get;}
-
+        public ICommand AddCinemaCommand { get; }
         public ICommand RemoveCinemaCommand { get; }
-
 
         public CinemaViewModel(ICinemaRepository cinemas)
         {
             _repo = cinemas;
 
+            Cinemas = new ObservableCollection<Cinema>(
+                _repo.GetAllCinemas()
+            );
 
-            Cinemas = new ObservableCollection<Cinema>(_repo.GetAllCinemas());
-
-
-
-            AddCinemaCommand = new RelayCommand(Parameter => AddCinema());
-
-            RemoveCinemaCommand = new RelayCommand(Parameter => RemoveCinema());
-
+            AddCinemaCommand = new RelayCommand(parameter => AddCinema());
+            RemoveCinemaCommand = new RelayCommand(parameter => RemoveCinema());
         }
 
         private string _name;
 
         public string Name
         {
-            get { return _name; }
-            set 
-            { 
+            get => _name;
+            set
+            {
                 _name = value;
                 OnPropertyChanged(nameof(Name));
             }
-            
+        }
+
+        private int _selectedNumberOfSale = 1;
+
+        public int SelectedNumberOfSale
+        {
+            get => _selectedNumberOfSale;
+            set
+            {
+                _selectedNumberOfSale = value;
+                OnPropertyChanged(nameof(SelectedNumberOfSale));
+            }
         }
 
         private ObservableCollection<Cinema> _cinemas;
@@ -60,44 +65,60 @@ namespace The_Movies.ViewModel
             }
         }
 
-
-
         private Cinema _selectedCinema;
 
         public Cinema SelectedCinema
         {
-            get { return _selectedCinema; }
-            set 
-            { 
+            get => _selectedCinema;
+            set
+            {
                 _selectedCinema = value;
                 OnPropertyChanged(nameof(SelectedCinema));
             }
         }
 
-
-
-
         private void AddCinema()
         {
-            Cinema cinema = new Cinema(Name);
+            // Opret en tom liste til salene
+            List<Sal> sale = new List<Sal>();
+
+            // Opret det antal sale som er valgt
+            for (int i = 1; i <= SelectedNumberOfSale; i++)
+            {
+                sale.Add(new Sal($"Sal {i}"));
+            }
+
+            // Opret biografen med salene
+            Cinema cinema = new Cinema(Name, sale);
+
+            // Gem biografen
             _repo.AddCinema(cinema);
 
+            // Tilføj til listen i UI
             Cinemas.Add(cinema);
 
+            // Ryd navn
+            Name = "";
+
+            // Sæt antal tilbage til 1
+            SelectedNumberOfSale = 1;
         }
 
         private void RemoveCinema()
         {
+            if (SelectedCinema == null)
+                return;
+
             _repo.RemoveCinema(SelectedCinema);
 
             Cinemas.Remove(SelectedCinema);
 
-
-
+            SelectedCinema = null;
         }
-
-
-
-
     }
 }
+
+
+
+    
+
